requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.3.1
django-dirtyfields==1.9.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environ==1.0
executing==2.0.1
frozenlist==1.4.1
icecream==2.1.3
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
pillow==10.3.0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.0
twilio==9.2.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
